+++ packages/resource-detector-azure/src/semconv.ts
@@ -49,6 +49,15 @@ export const ATTR_CLOUD_PROVIDER = 'cloud.provider' as const;
 export const ATTR_CLOUD_REGION = 'cloud.region' as const;
 
 /**
+ * Container name used by container runtime.
+ *
+ * @example opentelemetry-autoconf
+ *
+ * @experimental This attribute is experimental and is subject to breaking changes in minor releases of `@opentelemetry/semantic-conventions`.
+ */
+export const ATTR_CONTAINER_NAME = 'container.name' as const;
+
+/**
  * Name of the [deployment environment](https://wikipedia.org/wiki/Deployment_environment) (aka deployment tier).
  *
  * @example staging
@@ -66,6 +75,15 @@ export const ATTR_CLOUD_REGION = 'cloud.region' as const;
  */
 export const ATTR_DEPLOYMENT_ENVIRONMENT_NAME =
   'deployment.environment.name' as const;
+
+/**
+ * The name of the cluster.
+ *
+ * @example opentelemetry-cluster
+ *
+ * @experimental This attribute is experimental and is subject to breaking changes in minor releases of `@opentelemetry/semantic-conventions`.
+ */
+export const ATTR_K8S_CLUSTER_NAME = 'k8s.cluster.name' as const;
 
 /**
  * The execution environment ID as a string, that will be potentially reused for other invocations to the same function/function version.
@@ -182,6 +200,16 @@ export const CLOUD_PLATFORM_VALUE_AZURE_APP_SERVICE =
   'azure.app_service' as const;
 
 /**
+ * Enum value "azure.container_apps" for attribute {@link ATTR_CLOUD_PLATFORM}.
+ *
+ * Azure Container Apps
+ *
+ * @experimental This enum value is experimental and is subject to breaking changes in minor releases of `@opentelemetry/semantic-conventions`.
+ */
+export const CLOUD_PLATFORM_VALUE_AZURE_CONTAINER_APPS =
+  'azure.container_apps' as const;
+
+/**
  * Enum value "azure.functions" for attribute {@link ATTR_CLOUD_PLATFORM}.
  *
  * Azure Functions
@@ -198,6 +226,15 @@ export const CLOUD_PLATFORM_VALUE_AZURE_FUNCTIONS = 'azure.functions' as const;
  * @experimental This enum value is experimental and is subject to breaking changes in minor releases of `@opentelemetry/semantic-conventions`.
  */
 export const CLOUD_PLATFORM_VALUE_AZURE_VM = 'azure.vm' as const;
+
+/**
+ * Enum value "azure.aks" for attribute {@link ATTR_CLOUD_PLATFORM}.
+ *
+ * Azure Kubernetes Service
+ *
+ * @experimental This enum value is experimental and is subject to breaking changes in minor releases of `@opentelemetry/semantic-conventions`.
+ */
+export const CLOUD_PLATFORM_VALUE_AZURE_AKS = 'azure.aks' as const;
 
 /**
  * Enum value "azure" for attribute {@link ATTR_CLOUD_PROVIDER}.